+++ src/EPR.Calculator.API.UnitTests/Builder/CalcResultBuilderTests.cs
@@ -6,6 +6,7 @@ namespace EPR.Calculator.API.UnitTests.Builder
     using System;
     using EPR.Calculator.API.Builder;
     using EPR.Calculator.API.Builder.CommsCost;
+    using EPR.Calculator.API.Builder.Detail;
     using EPR.Calculator.API.Builder.LaDisposalCost;
     using EPR.Calculator.API.Builder.Lapcap;
     using EPR.Calculator.API.Builder.ParametersOther;
