+++ src/EPR.Calculator.API.UnitTests/Builder/CalcResultDetailBuilderTests.cs
@@ -1,7 +1,7 @@
 namespace EPR.Calculator.API.UnitTests.Builder
 {
     using System;
-    using EPR.Calculator.API.Builder;
+    using EPR.Calculator.API.Builder.Detail;
     using EPR.Calculator.API.Data;
     using EPR.Calculator.API.Dtos;
     using Microsoft.VisualStudio.TestTools.UnitTesting;
