+++ src/EPR.Calculator.API.UnitTests/CalResultsTests.cs
@@ -16,6 +16,7 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Moq;
 using EPR.Calculator.API.Builder.CommsCost;
+using EPR.Calculator.API.Builder.Detail;
 
 namespace EPR.Calculator.API.UnitTests
 {
